stageTwo/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.61.1
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.1
